image-dedup/pyproject.toml
[tool.poetry]
name = "image-dedup"
version = "0.1.0"
description = "Image deduplication pipeline with CLIP and PyQt"
authors = ["Senior Engineer <dev@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
# Фикс для PyTorch 2.1: запрещаем NumPy 2.0+
numpy = "<2.0.0"
# Фикс для SymPy: явно добавляем mpmath
mpmath = "^1.3.0"

open-clip-torch = "^2.23.0"
SQLAlchemy = "^2.0.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.0"}
pgvector = "^0.2.0"
typer = {extras = ["all"], version = "^0.12.3"}
PyQt6 = "^6.6.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
faiss-cpu = "^1.7.4"
networkx = "^3.2.1"
Pillow = "^10.0.0"
pyyaml = "^6.0"
tqdm = "^4.66.0"

# Явное указание источника для torch
torch = {version = "2.1.0", source = "torch_cu121"}
torchvision = {version = "0.16.0", source = "torch_cu121"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"

# Источник CUDA колес
[[tool.poetry.source]]
name = "torch_cu121"
url = "https://download.pytorch.org/whl/cu121"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.mypy]
ignore_missing_imports = true
